+++ app/src/main/AndroidManifest.xml
@@ -11,8 +11,8 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".PeripheralsList"
-            android:label="@string/title_activity_peripherals_list" >
+            android:name=".Peripherals"
+            android:label="@string/title_activity_peripherals" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -20,9 +20,9 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".Peripherals"
+                android:name=".PeripheralBroadcaster"
                 android:label="@string/title_activity_peripheral"
-                android:parentActivityName=".PeripheralsList"/>
+                android:parentActivityName=".Peripherals"/>
     </application>
 
 </manifest>
