+++ app/src/main/res/layout/activity_peripherals_list.xml
@@ -4,7 +4,7 @@
                 android:layout_height="match_parent"
                 android:paddingTop="@dimen/activity_vertical_margin"
                 android:paddingBottom="@dimen/activity_vertical_margin"
-                tools:context="io.github.webbluetoothcg.bletestperipheral.PeripheralsList">
+                tools:context="io.github.webbluetoothcg.bletestperipheral.Peripherals">
     <ListView
             android:id="@android:id/list"
             android:layout_width="match_parent"
