+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">BLE Test Peripheral</string>
-    <string name="title_activity_peripherals_list">Peripherals List</string>
+    <string name="title_activity_peripherals">Peripherals</string>
     <string name="title_activity_peripheral">Peripheral</string>
 
     <!-- Bluetooth -->
